Committing the report parsing, because what we've made might not actually be usable at all.
The data we're being given by OIT is extremely incomplete, so the administrators of the system are going to have to take a whole bunch of manual actions to get it fixed up. If we allow multiple imports of the report, all this manual reconciliation is going to be overwritten.